Fixing signed vs unsigned comparison warning
git-svn-id: svn://svn.code.sf.net/p/marathon/code/trunk@1957 a75d40c9-0969-4759-85e5-97cb871b7ed4

+++ Source_Files/Files/filetypes_macintosh.cpp
@@ -112,7 +112,7 @@ void initialize_typecodes()
 	Handle FTypHdl = GetResource('FTyp',128);
 	if (FTypHdl != NULL)
 	{
-		int FTHSize = GetHandleSize(FTypHdl);
+		size_t FTHSize = GetHandleSize(FTypHdl);
 		if (FTHSize <= sizeof(typecodes))
 		{
 			HLock(FTypHdl);
